requirements.txt
Pillow==9.4.0
Pillow==11.0.0
streamlit==1.40.0
